# contracts/openapi/api-specification.yaml
openapi: 3.0.3
info:
  title: BhoomiSetu AI API Specification
  description: REST API specification for BhoomiSetu AI - Early Detection of Land Acquisition Delays
  version: 1.0.0
servers:
  - url: http://localhost:8000/api
    description: Local Development Server
paths:
...
  /predictions/case/{case_id}:
    post:
      summary: Trigger on-demand AI risk prediction for a case
      parameters:
        - name: case_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: AI delay prediction result
  /predictions/case/{case_id}/latest:
    get:
      summary: Get latest AI risk prediction for a case
      parameters:
        - name: case_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Latest prediction
  /predictions/case/{case_id}/history:
    get:
      summary: Get prediction history for a case
      parameters:
        - name: case_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Prediction history timeline
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT